fix StatsTests rebase with master
