+++ src/OrleansTestingHost/Properties/AssemblyInfo.cs
@@ -22,3 +22,4 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("TesterInternal")]
 [assembly: InternalsVisibleTo("Tester.AzureUtils")]
 [assembly: InternalsVisibleTo("AWSUtils.Tests")]
+[assembly: InternalsVisibleTo("Tester.SQLUtils")]
